refactor(shared-docs): refactor docs-viewer and example-viewer into shared directory
Refactor into a shared directory in preperation for properly handling circular dep
