--- docs/components/example-viewer/BUILD.bazel
@@ -1,56 +1,0 @@
-load("//tools:defaults.bzl", "karma_web_test", "ng_module", "ts_library")
-load("@io_bazel_rules_sass//:defs.bzl", "sass_binary")
-
-ng_module(
-    name = "example-viewer",
-    srcs = [
-        "example-viewer.component.ts",
-    ],
-    assets = [
-        ":example-viewer.component.css",
-        "example-viewer.component.html",
-    ],
-    visibility = [
-        "//docs/components:__pkg__",
-        "//docs/components/docs-viewer:__pkg__",
-    ],
-    deps = [
-        "//docs/components/copy-source-code-button",
-        "//docs/interfaces",
-        "//docs/providers",
-        "@npm//@angular/cdk",
-        "@npm//@angular/common",
-        "@npm//@angular/core",
-        "@npm//@angular/material",
-    ],
-)
-
-sass_binary(
-    name = "style",
-    src = "example-viewer.component.scss",
-)
-
-ts_library(
-    name = "test_lib",
-    testonly = True,
-    srcs = glob(
-        ["*.spec.ts"],
-    ),
-    deps = [
-        ":example-viewer",
-        "//docs/components/copy-source-code-button",
-        "//docs/interfaces",
-        "//docs/providers",
-        "@npm//@angular/cdk",
-        "@npm//@angular/core",
-        "@npm//@angular/material",
-        "@npm//@angular/platform-browser",
-        "@npm//@angular/router",
-        "@npm//@types/jasmine",
-    ],
-)
-
-karma_web_test(
-    name = "test",
-    specs = [":test_lib"],
-)
