+++ docs/components/index.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-export * from './docs-viewer/docs-viewer.component';
+export * from './viewers/docs-viewer/docs-viewer.component';
 export * from './cookie-popup/cookie-popup.component';
 export * from './navigation-list/navigation-list.component';
 export * from './select/select.component';
