+++ docs/components/viewers/BUILD.bazel
@@ -2,12 +2,15 @@ load("//tools:defaults.bzl", "karma_web_test", "ng_module", "ts_library")
 load("@io_bazel_rules_sass//:defs.bzl", "sass_binary")
 
 ng_module(
-    name = "docs-viewer",
+    name = "viewers",
     srcs = [
-        "docs-viewer.component.ts",
+        "docs-viewer/docs-viewer.component.ts",
+        "example-viewer/example-viewer.component.ts",
     ],
     assets = [
-        ":docs-viewer.component.css",
+        ":docs-viewer/docs-viewer.component.css",
+        ":example-viewer/example-viewer.component.css",
+        "example-viewer/example-viewer.component.html",
     ],
     visibility = [
         "//docs/components:__pkg__",
@@ -15,19 +18,26 @@ ng_module(
     deps = [
         "//docs/components/breadcrumb",
         "//docs/components/copy-source-code-button",
-        "//docs/components/example-viewer",
         "//docs/components/table-of-contents",
+        "//docs/interfaces",
         "//docs/providers",
+        "@npm//@angular/cdk",
         "@npm//@angular/common",
         "@npm//@angular/core",
+        "@npm//@angular/material",
         "@npm//@angular/router",
         "@npm//@types/dom-view-transitions",
     ],
 )
 
 sass_binary(
-    name = "style",
-    src = "docs-viewer.component.scss",
+    name = "example-viewer-style",
+    src = "example-viewer/example-viewer.component.scss",
+)
+
+sass_binary(
+    name = "docs-viewer-style",
+    src = "docs-viewer/docs-viewer.component.scss",
     deps = [
         "//docs/styles",
     ],
@@ -37,19 +47,20 @@ ts_library(
     name = "test_lib",
     testonly = True,
     srcs = glob(
-        ["*.spec.ts"],
+        ["**/*.spec.ts"],
     ),
     deps = [
-        ":docs-viewer",
+        ":viewers",
         "//docs/components/breadcrumb",
         "//docs/components/copy-source-code-button",
-        "//docs/components/example-viewer",
         "//docs/components/icon",
         "//docs/components/table-of-contents",
         "//docs/interfaces",
         "//docs/providers",
         "//docs/services",
+        "@npm//@angular/cdk",
         "@npm//@angular/core",
+        "@npm//@angular/material",
         "@npm//@angular/platform-browser",
         "@npm//@angular/router",
         "@npm//@types/jasmine",
