+++ docs/components/viewers/example-viewer/example-viewer.component.html
